re-factor of component layout
