+++ src/Blur.js
@@ -1,0 +1,8 @@
+import React from "react";
+import "./App.css";
+
+const Blur = () => {
+  return <div className="blur"></div>;
+};
+
+export default Blur;
