+++ src/index.css
@@ -1,10 +1,10 @@
 body {
   margin: 0;
   font-family: "Roboto", sans-serif;
-  background-color: #2e4057;
+  background-color: white;
   display: flex;
   flex-direction: column;
   align-items: center;
   justify-content: center;
-  color: white;
+  color: black;
 }
